Fix db schema and refactor jpa repository

+++ src/main/java/kakaopay/pretask/moneyapi/api/dto/MoneyRecvResponse.java
@@ -3,12 +3,14 @@ package kakaopay.pretask.moneyapi.api.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class MoneyRecvResponse {
-	private long money;
+	private BigDecimal money;
 
-	public MoneyRecvResponse(long money) {
+	public MoneyRecvResponse(BigDecimal money) {
 		this.money = money;
 	}
 }
