+++ src/main/java/kakaopay/pretask/moneyapi/api/dto/MoneySpreadRequest.java
@@ -3,11 +3,13 @@ package kakaopay.pretask.moneyapi.api.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class MoneySpreadRequest {
 	private Long userId;
 	private String roomId;
-	private Long money;
-	private int headCount;
+	private BigDecimal money;
+	private Long headCount;
 }
